back/requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cloudinary==1.40.0
dj-database-url==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
PyJWT==2.8.0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.2
